+++ src/main/kotlin/wowsft/model/gameparams/ship/component/planes/HangarSetting.kt
@@ -5,5 +5,5 @@ class HangarSetting {
     var maxValue = 0
     var restoreAmount = 0
     var startValue = 0
-    var timeToRestore = 0f
+    var timeToRestore = 0.toDouble()
 }
